requirements.txt
backcall==0.1.0
colorama==0.4.3
decorator==4.4.2
dnspython==1.16.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jupyter-client==6.0.0
jupyter-core==4.6.3
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.2
pandas==1.0.3
parso==0.6.2
pickleshare==0.7.5
prompt-toolkit==3.0.4
protobuf==3.6.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pyzmq==19.0.0
six==1.14.0
SQLAlchemy==1.3.15
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.8
